requirements.txt
fastapi
uvicorn
openai
numpy
pillow
pytesseract
tiktoken
markdownify
beautifulsoup4
requests
